fix distutils.command.config thread safety
distutils.command.config has a number of methods that rely on generating
a temporary source file. The way this was done before was to name the
files something like `_configtest.c`, which ran into issues when the
methods were ran from multiple threads in the same directory. This
commit changes this behavior so that instead, the files will be named
something like `_configtest_<uuid4>.c`, where uuid4 is a thread safe
unique identifier.